FBTimer 0.91
Features:

Added the ability to synchronize the timer to loading a savestate. The default behavior is to load savestates saved using the script only; this is by design to prevent users from messing up their savestates, and to keep savestates used by the timer siloed away from the others.

Bugfixes:

Users are now no longer able to add or remove time from the timer while the timer is running.